Add tauri CI/CD

+++ build.py
@@ -91,6 +91,24 @@ def read_env(env_file='.env'):
     return os.environ
 
 
+def get_latest_changelog():
+    """Return the changelog entries for the latest (top-most) version in CHANGELOG.txt"""
+    changes = []
+    with open(CHANGELOG_FILE, encoding='utf-8') as _file:
+        found_version = False
+        for line in _file:
+            line = line.rstrip()
+            if not found_version:
+                # the first non-empty line after the title that isn't the title is the latest version
+                if line and line != 'Music Caster Changelog':
+                    found_version = True
+            elif line == '':
+                break
+            else:
+                changes.append(line)
+    return '\n'.join(changes)
+
+
 def add_new_changes(prev_changes: str):
     changes = set(prev_changes.split('\n'))
     with open(CHANGELOG_FILE, encoding='utf-8') as _file:
@@ -348,7 +366,23 @@ if __name__ == '__main__':
         action='store_true',
         help='if running in a CI do not prompt just fail',
     )
+    parser.add_argument(
+        '--changelog',
+        default=False,
+        action='store_true',
+        help="print the latest version's changelog and exit",
+    )
+    parser.add_argument(
+        '--target',
+        default=None,
+        help='Rust target triplet for the daemon artifact name '
+             '(default: {architecture}-pc-windows-msvc)',
+    )
     args = parser.parse_args()
+
+    if args.changelog:
+        print(get_latest_changelog())
+        sys.exit()
 
     if args.clean:
         shutil.rmtree(DIST_DIR, True)
@@ -584,10 +618,21 @@ if __name__ == '__main__':
                       'DOES NOT EXIST!')
                 dist_files_exist = False
 
-    if USING_TAURI_FRONTEND and platform.system() == 'Windows':
+    if USING_TAURI_FRONTEND:
         architecture = 'aarch64' if platform.machine() == 'ARM64' else 'x86_64'
-        shutil.copy2(daemon_dist, DIST_DIR / f'music-caster-daemon-{architecture}-pc-windows-msvc.exe')
-        print('copied daemon to target triplet')
+        if platform.system() == 'Windows':
+            target = args.target or f'{architecture}-pc-windows-msvc'
+            ext = 'exe'
+        else:
+            ext = daemon_dist.stem
+            if platform.system() == 'Linux':
+                target = args.target or f'{architecture}-unknown-linux-gnu'
+            else:
+                target = args.target
+                if target is None:
+                    raise Exception(f'target was not supplied for "{platform.system()}" platform')
+        shutil.copy2(daemon_dist, DIST_DIR / f'music-caster-daemon-{target}.{ext}')
+        print(f'copied daemon to target triplet {target}')
 
     if not args.skip_tests and dist_files_exist and not USING_TAURI_FRONTEND:
         try:
